\name{pcaGoPromoter.Rn.rn4}
					\docType{data}
					\alias{pcaGoPromoter.Rn.rn4}
					\alias{thresholds}
					\alias{promotersRefseqs}
					\title{The pcaGoPromoter.Rn.rn4 data set is for use with package pcaGoPromoters}
					\description{
					This data package contains two lists, thresholds and promotersRefseqs,
					which is used by the function \code{primo} in the software package
					'pcaGoPromoter'. See manual pages for function \code{primoData} in
					packages 'pcaGoPromoter' for description of the two variables.
					}
					\keyword{datasets}
					